changed back fetch
Вернула функцию)
